[team29][BE][제리 & 후] 2주차 두번째 PR

+++ BE/src/main/java/team29/hoorry/issuetracker/core/issue/dto/response/IssueMilestoneResponse.java
@@ -11,7 +11,7 @@ public class IssueMilestoneResponse {
 
 	private Long id;
 	private String title;
-	private Integer openIssueCount;
-	private Integer closedIssueCount;
+	private Long openIssueCount;
+	private Long closedIssueCount;
 
 }
